修改发票模块

+++ src/views/work/project/invociInfo.vue
@@ -5,30 +5,38 @@
           <a-row :gutter="24">
             <a-col :span="6">
               <a-form-item label="发票名称" >
-                <a-input placeholder="请输入发票名称" v-model="queryParam.invociName"></a-input>
+                <a-input placeholder="请输入发票名称" v-model="queryParam.invociName" maxLength="100"></a-input>
               </a-form-item>
             </a-col>
             <a-col :span="6">
               <a-form-item label="税号">
-                <a-input placeholder="请输入税号" v-model="queryParam.shuihao"></a-input>
+                <a-input placeholder="请输入税号" v-model="queryParam.shuihao" maxLength="20"></a-input>
               </a-form-item>
             </a-col>
-<!--
-            <a-form-item
-              :labelCol="labelCol"
-              :wrapperCol="wrapperCol"
-              label="开票时间"
-              hasFeedback >
-              <a-date-picker showTime format="YYYY-MM-DD"  v-model="queryParam.invociTime" />
-            </a-form-item>
--->
-
-
-
+            <a-col :span="6">
+              <a-form-item label="合同名称">
+                <a-input placeholder="请输入合同名称" v-model="queryParam.contractName" maxLength="50"></a-input>
+              </a-form-item>
+            </a-col>
+            <a-col :span="6">
+              <a-form-item label="公司名称">
+                <a-input placeholder="请输入公司名称" v-model="queryParam.companyName" maxLength="50"></a-input>
+              </a-form-item>
+            </a-col>
+          </a-row>
+          <a-row  :gutter="24" v-show="isShow">
+            <a-col :span="6">
+              <a-form-item label="开票时间">
+                <a-date-picker  v-model="queryParam.invociTime" />
+              </a-form-item>
+            </a-col>
+          </a-row>
+          <a-row  :gutter="24">
             <a-col :span="6"  >
               <span style="float: left;overflow: hidden;" class="table-page-search-submitButtons">
                 <a-button type="primary" @click="searchQuery" icon="search">查询</a-button>
                 <a-button type="primary" @click="searchReset" icon="reload" style="margin-left: 8px">重置</a-button>
+                <a-button type="primary" @click="superQuery" icon="filter" style="margin-left: 8px">高级查询</a-button>
               </span>
             </a-col>
           </a-row>
@@ -39,7 +47,7 @@
       <div class="table-operator">
         <a-button @click="handleAdd" type="primary" icon="plus">新增</a-button>
         <a-button type="primary" >
-          <a :href="'http://localhost:8080/jeecg-boot/renche/invoic/exportInvoic'" target="_blank" style="margin-left: 10px">导出</a>
+          <a :href="'http://localhost:8080/jeecg-boot/renche/invoci/exportInvoci'" target="_blank" style="margin-left: 10px">导出</a>
         </a-button>
 
 
@@ -83,24 +91,15 @@
 
             <a @click="handleEdit(record)">编辑</a>
             <a-divider type="vertical"/>
-            <a-dropdown>
-              <a class="ant-dropdown-link">更多 <a-icon type="down"/></a>
-              <a-menu slot="overlay">
-                <a-menu-item>
-                  <a-popconfirm title="确定删除吗?" @confirm="() => handleDelete(record.invociId )">
-                    <a>删除</a>
-                  </a-popconfirm>
-                </a-menu-item>
-              </a-menu>
-            </a-dropdown>
+            <a @click="handleDelete(record.invociId )">删除</a>
           </span>
         </a-table>
       </div>
       <!-- table区域-end -->
 
       <!-- 表单区域 -->
-        <add-invoic-info ref="addInvoicInfo" @ok="modalFormOk"></add-invoic-info>
-      <invoic-info-file-detail ref="invoicInfoFileDetail" @ok="modalFormOk"></invoic-info-file-detail>
+        <add-invoci-info ref="addInvociInfo" @ok="modalFormOk"></add-invoci-info>
+      <invoci-info-file-detail ref="invociInfoFileDetail" @ok="modalFormOk"></invoci-info-file-detail>
 
 
     </a-card>
@@ -109,23 +108,23 @@
 
 <script>
     import ARow from "ant-design-vue/es/grid/Row";
-    import addInvoicInfo from './modules/addInvoicInfo';
+    import AddInvociInfo from './modules/AddInvociInfo';
     import {filterObj} from '@/utils/util';
-    import invoicInfoFileDetail from "./modules/invoicInfoFileDetail";
+    import InvociInfoFileDetail from "./modules/InvociInfoFileDetail";
     import {deleteAction, getAction, postAction} from '@/api/manage';
 
     export default {
       name: "invoicInfo",
       components: {
         ARow,
-        addInvoicInfo,
-        invoicInfoFileDetail,
+        AddInvociInfo,
+        InvociInfoFileDetail,
 
       },
       data() {
         return{
           description: '招标管理页面',
-
+          isShow:false,
           // 查询条件
           queryParam: {},
           // 表头
@@ -141,12 +140,17 @@
               }
             },
             {
+              title: '合同名称',
+              align: "center",
+              dataIndex: 'contractName',
+            },
+            {
               title: '发票名称',
               align: "center",
               dataIndex: 'invociName',
             },
             {
-              title: '开票时间',
+              title: '开票日期',
               align: "center",
               dataIndex: 'invociTime',
             },
@@ -156,41 +160,20 @@
               dataIndex: 'content',
             },
             {
+              title: '公司名称',
+              align: "center",
+              dataIndex: 'companyName',
+            },
+            {
               title: '税号',
               align: "center",
               dataIndex: 'shuihao',
             },
-            {
-              title: '单位地址',
-              align: "center",
-              dataIndex: 'address',
-
-            },
-            {
-              title: '电话号码',
-              align: "center",
-              dataIndex: 'tel',
-            },
-            {
-              title: '开户银行',
-              align: "center",
-              dataIndex: 'bank',
-            },
-            {
-              title: '银行账户',
-              align: "center",
-              dataIndex: 'bankNo',
-            },
-            {
-              title: '创建时间',
-              align: "center",
-              dataIndex: 'createTime',
-            },
-          /*  {
+           {
               title: '附件',
               align: "center",
-              dataIndex: 'fileRelNum',
-            },*/
+              dataIndex: 'fileCount',
+            },
             {
               title: '操作',
               dataIndex: 'action',
@@ -205,8 +188,8 @@
           // 分页参数
           ipagination: {
             current: 1,
-            pageSize: 10,
-            pageSizeOptions: ['10', '20', '30'],
+            pageSize: 30,
+            pageSizeOptions: ['20', '30', '40'],
             showTotal: (total, range) => {
               return range[0] + "-" + range[1] + " 共" + total + "条"
             },
@@ -222,9 +205,10 @@
           selectedRowKeys: [],
           selectedRows: [],
           url: {
-            list: "/renche/invoic/qryInvoicList",
-            delete: "/renche/invoic/delete",
-            deleteBatch: "/renche/invoic/deleteBat",
+            list: "/renche/invoci/qryInvociList",
+            delete: "/renche/invoci/delete",
+            deleteBatch: "/renche/invoci/deleteBat",
+            filelist: "/renche/purchase/fileList",
           },
         }
       },
@@ -265,7 +249,11 @@
           this.loadData(1);
         },
         superQuery() {
-          this.$refs.superQueryModal.show();
+          if(this.isShow){
+            this.isShow = false;
+          }else{
+            this.isShow = true;
+          }
         },
         searchReset() {
           var that = this;
@@ -313,29 +301,38 @@
         },
         handleDelete: function (id) {
           var that = this;
-          deleteAction(that.url.delete, {id: id}).then((res) => {
-            if (res.success) {
-              that.$message.success(res.message);
-              that.loadData();
-            } else {
-              that.$message.warning(res.message);
+          this.$confirm({
+            title: "是否确认删除该数据?",
+            onOk: function () {
+              deleteAction(that.url.delete, {id: id}).then((res) => {
+                if (res.success) {
+                  that.$message.success(res.message);
+                  that.loadData();
+                } else {
+                  that.$message.warning(res.message);
+                }
+              });
             }
           });
         },
-        handleEdit: function (record) {
-          this.$refs.addInvoicInfo.edit(record);
-          this.$refs.addInvoicInfo.title = "编辑";
+        async handleEdit (record) {
+          if(record.fileRelId != null || record.fileRelId != "" || record.fileRelId != undefined) {
+            let {result} = await getAction(this.url.filelist, {fileRelId: record.fileRelId});
+            record.filelist = result.list;
+          }
+          this.$refs.addInvociInfo.edit(record);
+          this.$refs.addInvociInfo.title = "编辑";
         },
 
         fileDeteil:function(record){
           console.log(record);
-          this.$refs.invoicInfoFileDetail.fileLoad(record);
-          this.$refs.invoicInfoFileDetail.title = "附件";
+          this.$refs.invociInfoFileDetail.fileLoad(record);
+          this.$refs.invociInfoFileDetail.title = "附件";
         },
 
         handleAdd: function () {
-          this.$refs.addInvoicInfo.add();
-          this.$refs.addInvoicInfo.title = "新增";
+          this.$refs.addInvociInfo.add();
+          this.$refs.addInvociInfo.title = "新增";
         },
         handleTableChange(pagination, filters, sorter) {
           //分页、排序、筛选变化时触发
@@ -355,14 +352,9 @@
         },
         modalFormOk() {
           // 新增/修改 成功时，重载列表
-          this.searchQuery();
-
-
+          this.loadData(1);
         }
       },
-   /*   mounted() {
-        /!*this.timer = setInterval(this.loadData, 10000);*!/
-      },*/
     }
 
 </script>
